pyproject.toml
[tool.black]
line-length = 120
exclude = '''
/(
    \.git
  | venv
  | \.venv
  | env
  | \.env
  | _build
  | build
  | dist
)/
'''

[tool.poetry]
name = "coordinate-projector"
version = "0.0.11"
description = "Project points from one projection to another using pyproj"
license = "MIT"
authors = ["Helge Smebye", "Jostein Leira <jostein@leira.net>"]
maintainers = ["Jostein Leira <jostein@leira.net>"]
readme = "README.md"
homepage = "https://github.com/norwegian-geotechnical-institute/coordinate-projector"
repository = "https://github.com/norwegian-geotechnical-institute/coordinate-projector"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: GIS",
]
packages = [
    { include = "coordinate_projector", from="src" },
]


[tool.ruff]
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.poetry.dependencies]
python = ">=3.10,<4"
pyproj = "^3.6.1"
python-dateutil = "^2.8.2"
types-python-dateutil = "^2.8.19"
timezonefinder = "*"


[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
mypy = "*"
bandit = "*"
safety = "*"
pip-licenses = "*"
ruff = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
